requirements.txt
# Invoice Processing System - Requirements
# Install with: pip install -r requirements.txt

# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# OCR
pytesseract==0.3.10
Pillow==10.1.0
opencv-python==4.8.1.78

# PDF processing
PyMuPDF==1.23.8
pdf2image==1.16.3

# AI/LLM
openai==1.3.7
python-dotenv==1.0.0

# Excel
openpyxl==3.1.2
pandas==2.1.4

# Database (optional)
sqlalchemy==2.0.23

# Utilities
aiofiles==23.2.1
jinja2==3.1.2
